+++ src/domains/wallet/mod.rs
@@ -1,10 +1,12 @@
 mod entities;
+mod user_wallet_handler;
 mod wallet_handler;
 mod wallet_repository;
 mod wallet_service;
 mod wallet_use_cases;
 
 pub use entities::*;
+pub use user_wallet_handler::*;
 pub use wallet_handler::*;
 pub use wallet_repository::*;
 pub use wallet_service::*;
